+++ app/src/main/java/rkr/simplekeyboard/inputmethod/latin/settings/Settings.java
@@ -54,6 +54,7 @@ public final class Settings extends BroadcastReceiver implements SharedPreferenc
     public static final String PREF_SOUND_ON = "sound_on";
     public static final String PREF_POPUP_ON = "popup_on";
     public static final String PREF_SHOW_LANGUAGE_SWITCH_KEY = "pref_show_language_switch_key";
+    public static final String PREF_SHOW_EMOJI_KEY = "pref_show_emoji_key";
     public static final String PREF_USE_ON_SCREEN = "pref_use_on_screen";
     public static final String PREF_ENABLE_IME_SWITCH = "pref_enable_ime_switch";
     public static final String PREF_ENABLED_SUBTYPES = "pref_enabled_subtypes";
@@ -269,6 +270,10 @@ public final class Settings extends BroadcastReceiver implements SharedPreferenc
 
     public static boolean readShowLanguageSwitchKey(final SharedPreferences prefs) {
         return prefs.getBoolean(PREF_SHOW_LANGUAGE_SWITCH_KEY, true);
+    }
+
+    public static boolean readShowEmojiKey(final SharedPreferences prefs) {
+        return prefs.getBoolean(PREF_SHOW_EMOJI_KEY, true);
     }
 
     public static boolean readUseOnScreenKeyboard(final SharedPreferences prefs) {