+++ app/src/main/java/rkr/simplekeyboard/inputmethod/latin/settings/SettingsValues.java
@@ -47,6 +47,7 @@ public class SettingsValues {
     public final boolean mKeyPreviewPopupOn;
     public final boolean mUseOnScreen;
     public final boolean mShowsLanguageSwitchKey;
+    public final boolean mShowEmojiKey;
     public final boolean mImeSwitchEnabled;
     public final int mKeyLongpressTimeout;
     public final boolean mShowSpecialChars;
@@ -84,6 +85,7 @@ public class SettingsValues {
         mKeyPreviewPopupOn = Settings.readKeyPreviewPopupEnabled(prefs, res);
         mUseOnScreen = Settings.readUseOnScreenKeyboard(prefs);
         mShowsLanguageSwitchKey = Settings.readShowLanguageSwitchKey(prefs);
+        mShowEmojiKey = Settings.readShowEmojiKey(prefs);
         mImeSwitchEnabled = Settings.readEnableImeSwitch(prefs);
         mHasHardwareKeyboard = Settings.readHasHardwareKeyboard(res.getConfiguration());